Some further project cleanup.
* Moved Angle, Vector3 and Twist classes into separate header files and removed loam_types.h again.
